+++ lib/td_cache/concept_cache.ex
@@ -17,6 +17,7 @@ defmodule TdCache.ConceptCache do
   @keys "business_concept:keys"
   @active_ids "business_concept:ids:active"
   @inactive_ids "business_concept:ids:inactive"
+  @confidential_ids "business_concept:ids:confidential"
 
   ## Client API
 
@@ -56,6 +57,13 @@ defmodule TdCache.ConceptCache do
   end
 
   @doc """
+  Updates cache entries for confidential ids.
+  """
+  def put_confidential_ids(ids) do
+    GenServer.call(__MODULE__, {:confidential_ids, ids})
+  end
+
+  @doc """
   Reads active concept ids (as a list of strings) from cache.
   """
   def active_ids do
@@ -63,10 +71,24 @@ defmodule TdCache.ConceptCache do
   end
 
   @doc """
+  Reads confidential concept ids (as a list of strings) from cache.
+  """
+  def confidential_ids do
+    GenServer.call(__MODULE__, :confidential_ids)
+  end
+
+  @doc """
   Deletes cache entries relating to a given concept id.
   """
   def delete(id) do
     GenServer.call(__MODULE__, {:delete, id})
+  end
+
+  @doc """
+  Verifies if id is member of confidential ids set.
+  """
+  def member_confidential_ids(id) do
+    GenServer.call(__MODULE__, {:member, :confidential_ids, id})
   end
 
   ## Callbacks
@@ -130,9 +152,21 @@ defmodule TdCache.ConceptCache do
   end
 
   @impl true
+  def handle_call(:confidential_ids, _from, state) do
+    ids = read_confidential_ids()
+    {:reply, {:ok, ids}, state}
+  end
+
+  @impl true
   def handle_call({:delete, id}, _from, state) do
     reply = delete_concept(id)
     {:reply, reply, state}
+  end
+
+  @impl true
+  def handle_call({:member, :confidential_ids, id}, _from, state) do
+    reply = is_member_confidential_ids?(id)
+    {:reply, {:ok, reply}, state}
   end
 
   @impl true
@@ -141,9 +175,16 @@ defmodule TdCache.ConceptCache do
     {:reply, reply, state}
   end
 
+  @impl true
+  def handle_call({:confidential_ids, ids}, _from, state) do
+    reply = update_confidential_ids(ids)
+    {:reply, reply, state}
+  end
+
   ## Private functions
 
   @props [:name, :domain_id, :business_concept_version_id, :current_version]
+  @confidential "Si"
 
   defp get_cache(key, fun) do
     ConCache.get_or_store(:concepts, key, fn -> fun.() end)
@@ -190,11 +231,12 @@ defmodule TdCache.ConceptCache do
       ["DEL", "business_concept:#{id}"],
       ["SREM", @keys, "business_concept:#{id}"],
       ["SADD", @inactive_ids, id],
-      ["SREM", @active_ids, id]
+      ["SREM", @active_ids, id],
+      ["SREM", @confidential_ids, id]
     ]
 
     results = Redix.transaction_pipeline!(commands)
-    [_, _, inactivated, _] = results
+    [_, _, inactivated, _, _] = results
 
     unless inactivated == 0 do
       publish_event("remove_concepts", id)
@@ -203,16 +245,22 @@ defmodule TdCache.ConceptCache do
     {:ok, results}
   end
 
+  defp is_member_confidential_ids?(id) do
+    ["SISMEMBER", @confidential_ids, id] |> Redix.command!()
+  end
+
   defp put_concept(%{id: id} = concept) do
-    commands = [
-      ["HMSET", "business_concept:#{id}", Map.take(concept, @props)],
-      ["SADD", @keys, "business_concept:#{id}"],
-      ["SREM", @inactive_ids, id],
-      ["SADD", @active_ids, id]
-    ]
+    commands =
+      [
+        ["HMSET", "business_concept:#{id}", Map.take(concept, @props)],
+        ["SADD", @keys, "business_concept:#{id}"],
+        ["SREM", @inactive_ids, id],
+        ["SADD", @active_ids, id]
+      ]
+      |> confidential_ids_command(concept)
 
     results = Redix.transaction_pipeline!(commands)
-    [_, _, activated, _] = results
+    [_, _, activated, _, _] = results
 
     unless activated == 0 do
       publish_event("restore_concepts", id)
@@ -223,6 +271,11 @@ defmodule TdCache.ConceptCache do
 
   defp read_active_ids do
     ["SMEMBERS", @active_ids]
+    |> Redix.command!()
+  end
+
+  def read_confidential_ids do
+    ["SMEMBERS", @confidential_ids]
     |> Redix.command!()
   end
 
@@ -245,6 +298,33 @@ defmodule TdCache.ConceptCache do
     publish_event("restore_concepts", restored_ids)
     publish_event("remove_concepts", removed_ids)
     {:ok, results}
+  end
+
+  defp update_confidential_ids(ids) do
+    commands = [
+      ["DEL", @confidential_ids],
+      ["SADD", @confidential_ids] ++ ids,
+      ["SMEMBERS", @confidential_ids]
+    ]
+
+    results = Redix.transaction_pipeline!(commands)
+    [_, _, confidential_ids] = results
+    publish_event("confidential_concepts", confidential_ids)
+    {:ok, results}
+  end
+
+  defp confidential_ids_command(commands, %{id: id, content: content}) do
+    confidential_command =
+      case Map.get(content, "_confidential") do
+        @confidential -> ["SADD", @confidential_ids, id]
+        _ -> ["SREM", @confidential_ids, id]
+      end
+
+    commands ++ [confidential_command]
+  end
+
+  defp confidential_ids_command(commands, %{id: id}) do
+    commands ++ [["SREM", @confidential_ids, id]]
   end
 
   defp publish_event(_, []), do: :ok